refactor tile coverage code into footprint module
